# [proof of concept] concurrent stackwalking