--- a/minidump-processor/Cargo.toml
+++ b/minidump-processor/Cargo.toml
@@ -37,6 +37,7 @@ scroll = "0.11.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 thiserror = "1.0.30"
+futures-util = "0.3.21"
 
 [dev-dependencies]
 doc-comment = "0.3.3"