This is a proof-of-concept for:

1. making breakpad-symbols more properly async by reworking the streaming parser to be more stateful and not in charge of the http Response and only* using tokio::fs.
2. making minidump-processor `join` all the stackwalks

Commit 1 is the bulk of the work, Commit 2 is just a quick and dirty hack.

The net result is that all stackwalking is now concurrent, but not parallel. hot stackwalks take about the same amount of time, because file i/o is ultimately blocking, while cold (http) stackwalks on firefox take about half the time from being able to download things in parallel.

Subsequent comments will include some notes/numbers.