@Gankra 2022-07-01
Symbolizer is ultimately pretty self-contained when it comes to acquiring a SymbolFile for a module, and that's where all the hard i/o happens, so I expect the next lowest-hanging fruit is to `spawn` actual parallel tasks somewhere around the `locate_symbols` boundary. This would require wrapping a few pieces of state in an Arc, which should actually be minimally intrusive, since both HttpSymbolSupplier and SimpleSymbolSupplier already own all their state and work with shared references. The only mutable state *any* of them has is the cached_file_paths in http for downloading native binaries, which is already setup for parallelism (wrapped in a Mutex).