@Gankra 2022-07-01
I'm not sure if/when I'll have more time to work on this, I just had an epiphany on how to disentangle the parser and wanted to put this up.