@Gankra 2022-07-01
possibly of interest to @Swatinem and @mstange who are thinking about perf/parsing. The new approach doesn't actually touch the "core" parser at all, so it's still completely swappable.