@Gankra 2022-07-01
new commit: very naively wraps the SymbolSuppliers and Module in Arcs and spawns a task when starting those queries. In principle this makes Symbol loading/downloading/parsing truly parallel, although obviously the subtask for each thread is still totally blocked on this. I also still don't join on potential parallel paths within the symbol parser (i.e. trying every url in parallel).

# Benchmark Summary

* Hot (local): 1573ms -> 1462ms
* Cold (http): 3433ms -> 3270ms
* memory: 248MB -> 267MB

Some modest perf wins in exchange for some modest memory losses. Iterating past this point merits some profiling.

<details>

<summary> benchmark results </summary>

# New Hot (local)

```
benchmark results (ms):
  1472, 1450, 1451, 1474, 1466, 
average runtime: 00m:01s:462ms (1462ms)
median runtime: 00m:01s:466ms (1466ms)
min runtime: 00m:01s:450ms (1450ms)
max runtime: 00m:01s:474ms (1474ms)
```

# New Cold (http)

```
benchmark results (ms):
  3395, 3271, 3203, 3261, 3222, 
average runtime: 00m:03s:270ms (3270ms)
median runtime: 00m:03s:261ms (3261ms)
min runtime: 00m:03s:203ms (3203ms)
max runtime: 00m:03s:395ms (3395ms)
```



</details>